requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.1
Django==5.0.3
django-crispy-forms==2.1
pillow==10.3.0
sqlparse==0.4.4
tzdata==2024.1